+++ tests/test_builtin.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   test_builtin.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: krocha-h <krocha-h@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/06/01 15:10:23 by krocha-h          #+#    #+#             */
+/*   Updated: 2024/06/01 15:14:13 by krocha-h         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "test.h"
+
+int	test_pwd()
+{
+	// char	*path;
+	
+	// path = "/home/haydkelly/Galaxia/minishell/tests";
+	pwd();
+	return (1);
+}
